Move operation timeouts into core deadlines
Constraint: withTimeout enforces ToolContext deadlines and belongs beside the shared Budget primitive.
Rejected: Keep withTimeout in src/utils | deadline enforcement is core execution infrastructure, not a miscellaneous helper.
Confidence: high
Scope-risk: moderate
Directive: Import deadline-aware promise racing from src/core/deadline/with-timeout.ts.
Tested: npm test -- --runTestsByPath tests/core/deadline/with-timeout.test.ts tests/tools/act.test.ts tests/tools/interact.coordinate.test.ts tests/tools/interact-opened-tabs.test.ts tests/tools/performance-metrics-contract-facts.test.ts tests/dom/dom-serializer-timeout.test.ts tests/core/screenshot/guards.test.ts --runInBand; npm run build; npm run lint; npm run lint:tier; npm run lint:repo-structure; git diff --check
Not-tested: GitHub PR checks not yet run
