+++ src/tools/oc-vitals.ts
@@ -4,7 +4,7 @@ import { MCPServer } from '../mcp-server';
 import { MCPToolDefinition, MCPResult, ToolHandler } from '../types/mcp';
 import { TOOL_ANNOTATIONS } from '../types/tool-annotations';
 import { getSessionManager } from '../session-manager';
-import { withTimeout } from '../utils/with-timeout';
+import { withTimeout } from '../core/deadline/with-timeout';
 import { collectWebVitalsInPage, normalizeWebVitals, RawWebVitals } from '../core/perf/web-vitals-collector';
 
 const definition: MCPToolDefinition = {
