+++ src/tools/performance-metrics.ts
@@ -6,7 +6,7 @@ import { MCPServer } from '../mcp-server';
 import { MCPToolDefinition, MCPResult, ToolContext, ToolHandler } from '../types/mcp';
 import { TOOL_ANNOTATIONS } from '../types/tool-annotations';
 import { getSessionManager } from '../session-manager';
-import { withTimeout } from '../utils/with-timeout';
+import { withTimeout } from '../core/deadline/with-timeout';
 import { buildPerformanceContractFacts } from '../contracts/contract-facts';
 import { redactSecretStringWithMetadata } from '../core/secrets/redactor';
